+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-06-23 16:22:51
-// Commit: style(perf): harmonize light and dark theme palette for perf
+// Updated at: 2026-06-24 15:04:34
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '2233';
-  final String lastMessage = 'style(perf): harmonize light and dark theme palette for perf';
-  final DateTime timestamp = DateTime.parse('2026-06-23T16:22:51+05:00');
+  final String commitRef = '2262';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-06-24T15:04:34+05:00');
 }